+++ src/App/components/index.js
@@ -1,6 +1,6 @@
 export { Container } from 'App/components/containers/container.js';
-export { Content } from 'App/components/containers/content.js';
+export { Content } from 'App/components/containers/content';
 export { Error } from 'App/components/containers/base-alert.js';
-export { Icon } from 'App/components/icon/icon.js';
+export { Icon } from 'App/components/icon';
 export { LoadingResult } from 'App/components/containers/loading-result.js';
 export { Message } from 'App/components/containers/message.js';
