--- src/App/pages/search/search.module.css
@@ -1,64 +1,0 @@
-.typography {
-  color: var(--low-contrast-text);
-  font-size: var(--mantine-font-size-sm);
-
-  & ul,
-  & ol {
-    & li {
-      font-size: var(--mantine-font-size-sm);
-    }
-  }
-
-  & blockquote {
-    font-size: var(--mantine-font-size-sm);
-    border-top-right-radius: var(--mantine-radius-sm);
-    border-bottom-right-radius: var(--mantine-radius-sm);
-    color: var(--low-contrast-text);
-    border-left: rem(5) solid var(--mantine-color-default-border);
-
-    & cite {
-      display: block;
-      font-size: var(--mantine-font-size-sm);
-      margin-top: var(--mantine-spacing-xs);
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-  }
-}
-
-.item {
-  border: 1px solid var(--subtle-border-and-separator-blue);
-  background-color: var(--app-background);
-  border-radius: var(--mantine-radius-xl);
-  padding: var(--mantine-spacing-sm);
-  margin-bottom: var(--mantine-spacing-xs);
-
-  @mixin hover {
-    border-color: var(--hovered-ui-element-border-blue);
-  }
-}
-
-.resulTitle {
-  color: var(--high-contrast-text);
-}
-
-.externalIcon {
-  display: none;
-}
-
-.resulTitle:hover .externalIcon {
-  display: revert;
-}
-
-.spoilerControl {
-  font-size: var(--mantine-font-size-sm);
-  color: var(--high-contrast-text);
-  background-color: var(--ui-element-background);
-  text-align: center;
-  width: 100%;
-
-  @mixin hover {
-    text-decoration: none;
-    background-color: var(--hovered-ui-element-background);
-  }
-}
