+++ src/App/pages/search/typography/index.module.css
@@ -1,0 +1,27 @@
+.typography {
+  color: var(--low-contrast-text);
+  font-size: var(--mantine-font-size-sm);
+
+  & ul,
+  & ol {
+    & li {
+      font-size: var(--mantine-font-size-sm);
+    }
+  }
+
+  & blockquote {
+    font-size: var(--mantine-font-size-sm);
+    border-top-right-radius: var(--mantine-radius-sm);
+    border-bottom-right-radius: var(--mantine-radius-sm);
+    color: var(--low-contrast-text);
+    border-left: rem(5) solid var(--mantine-color-default-border);
+
+    & cite {
+      display: block;
+      font-size: var(--mantine-font-size-sm);
+      margin-top: var(--mantine-spacing-xs);
+      overflow: hidden;
+      text-overflow: ellipsis;
+    }
+  }
+}
